Fix issues with the struture of lib/app
