+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "rpa-em-assembly-webapp",
+  "name": "rpa-dg-docassembly-demo",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
@@ -29,7 +29,7 @@
     "@ngx-formly/core": "^4.8.2",
     "core-js": "^2.5.4",
     "govuk-frontend": "^2.5.1",
-    "jui-formly-lib": "0.0.3",
+    "govuk-formly-templates": "^0.1.6-RC",
     "rxjs": "~6.3.3",
     "tslib": "^1.9.0",
     "zone.js": "~0.8.26"
