+++ projects/rpa-dg-docassembly-webcomponent/src/public_api.ts
@@ -2,4 +2,4 @@
  * Public API Surface of govuk-formly-templates
  */
 
-export * from './lib/app.module';
+export * from './lib/assembly.module';
